refactor: 패키지명 수정 및 스파크 코드 제거
