+++ src/main/java/wooteco/chess/domain/move/direction/LeftDownStrategy.java
@@ -1,4 +1,4 @@
-package wooteco.chess.domain.strategy.move.direction;
+package wooteco.chess.domain.move.direction;
 
 import wooteco.chess.domain.position.Position;
 import wooteco.chess.domain.position.Positions;
@@ -6,13 +6,13 @@ import wooteco.chess.domain.position.Positions;
 import java.util.ArrayList;
 import java.util.List;
 
-public class RightDownStrategy implements DirectionStrategy {
+public class LeftDownStrategy implements DirectionStrategy {
     @Override
     public List<Position> findPath(Position source, Position target) {
         List<Position> path = new ArrayList<>();
 
-        for (int i = target.getFile() - 1; i > source.getFile(); i--) {
-            path.add(Positions.of(i, target.getRank() + (target.getFile() - i)));
+        for (int i = target.getFile() + 1; i < source.getFile(); i++) {
+            path.add(Positions.of(i, target.getRank() - (target.getFile() - i)));
         }
 
         return path;
