+++ src/main/java/wooteco/chess/domain/move/direction/RightDownStrategy.java
@@ -1,4 +1,4 @@
-package wooteco.chess.domain.strategy.move.direction;
+package wooteco.chess.domain.move.direction;
 
 import wooteco.chess.domain.position.Position;
 import wooteco.chess.domain.position.Positions;
@@ -6,14 +6,15 @@ import wooteco.chess.domain.position.Positions;
 import java.util.ArrayList;
 import java.util.List;
 
-public class RightStrategy implements DirectionStrategy {
+public class RightDownStrategy implements DirectionStrategy {
     @Override
     public List<Position> findPath(Position source, Position target) {
         List<Position> path = new ArrayList<>();
 
         for (int i = target.getFile() - 1; i > source.getFile(); i--) {
-            path.add(Positions.of(i, source.getRank()));
+            path.add(Positions.of(i, target.getRank() + (target.getFile() - i)));
         }
+
         return path;
     }
 }
