+++ src/main/java/wooteco/chess/domain/piece/PieceType.java
@@ -2,13 +2,13 @@ package wooteco.chess.domain.piece;
 
 import wooteco.chess.domain.board.Board;
 import wooteco.chess.domain.position.Position;
-import wooteco.chess.domain.strategy.move.BishopMoveStrategy;
-import wooteco.chess.domain.strategy.move.KingMoveStrategy;
-import wooteco.chess.domain.strategy.move.KnightMoveStrategy;
-import wooteco.chess.domain.strategy.move.MoveStrategy;
-import wooteco.chess.domain.strategy.move.PawnMoveStrategy;
-import wooteco.chess.domain.strategy.move.QueenMoveStrategy;
-import wooteco.chess.domain.strategy.move.RookMoveStrategy;
+import wooteco.chess.domain.move.BishopMoveStrategy;
+import wooteco.chess.domain.move.KingMoveStrategy;
+import wooteco.chess.domain.move.KnightMoveStrategy;
+import wooteco.chess.domain.move.MoveStrategy;
+import wooteco.chess.domain.move.PawnMoveStrategy;
+import wooteco.chess.domain.move.QueenMoveStrategy;
+import wooteco.chess.domain.move.RookMoveStrategy;
 
 public enum PieceType {
     KING("k", 0, new KingMoveStrategy()),
