+++ src/main/java/wooteco/chess/dto/Commands.java
@@ -1,7 +1,6 @@
 package wooteco.chess.dto;
 
 import org.springframework.data.annotation.Id;
-
 public class Commands {
     @Id
     private Long id;
